requirements.txt
# telegram-intel/requirements.txt

# --- Core ---
python-dotenv # Для загрузки .env файлов

# --- Telegram API ---
telethon==1.35.0

# --- Web Backend (FastAPI) ---
fastapi==0.110.0
uvicorn[standard]>=0.20.0 # ASGI сервер для FastAPI
websockets>=10.0 # Для поддержки WebSocket в FastAPI

# --- Database (PostgreSQL + SQLAlchemy + Alembic) ---
sqlalchemy==2.0.27
# Используем asyncpg т.к. FastAPI и SQLAlchemy 2.0 поддерживают async
asyncpg>=0.25.0 # Асинхронный драйвер PostgreSQL
alembic>=1.10.0 # Для миграций БД
psycopg2-binary>=2.9.0 # Добавляем и синхронный драйвер, т.к. Alembic по умолчанию может его использовать, или для других скриптов

# --- Task Queues (Celery + Redis) ---
celery==5.3.6 # Указываем последнюю стабильную версию в ветке 5.3.x
redis>=4.5.0 # Клиент Redis для Celery broker/backend

# --- Authentication ---
passlib[bcrypt]>=1.7.4 # Для хэширования паролей

# --- Data Validation (часто используется с FastAPI) ---
pydantic>=2.0.0 # Уже является зависимостью FastAPI, но явно укажем
email-validator>=1.1.3 # Для валидации email в моделях Pydantic

# --- Code Quality Tools (можно вынести в dev-requirements.txt) ---
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
# ... другие зависимости ...
PyJWT>=2.0.0